requirements.txt
numpy==1.20.1
pandas==1.2.4
scikit-learn==0.24.1
pyyaml==5.4.1
matplotlib==3.3.4
dvc==2.4.1
